Fixed schema migrations on MySQL where multiple schemas are under management by gomigrate.
The 'table exists' check for the migrations table wasn't qualified with the current
database, resulting in a false assumption that the migrations table exists in the current
database when in fact it is associated with a different database on the same server.  This
ultimately results in the migration failing for the current database with an error akin to
'migrations table not found'.